requirements.txt
requests
azure-storage-blob
black